fix: macro and react library should not mix
