+++ package.json
@@ -11,6 +11,7 @@
     "@fontsource/poppins": "5.0.8",
     "@jonkoops/matomo-tracker-react": "0.7.0",
     "@lingui/react": "4.7.0",
+    "@lingui/macro": "4.7.0",
     "@mui/icons-material": "5.15.7",
     "@mui/lab": "5.0.0-alpha.163",
     "@mui/material": "5.15.7",
@@ -64,7 +65,6 @@
     "@graphql-codegen/typescript-react-apollo": "4.1.0",
     "@graphql-codegen/visitor-plugin-common": "4.0.1",
     "@lingui/cli": "4.7.0",
-    "@lingui/macro": "4.7.0",
     "@lingui/vite-plugin": "4.7.0",
     "@types/chart.js": "2.9.41",
     "@types/classnames": "2.3.0",
