+++ src/components/AppBar/LandingPageAppBar.tsx
@@ -11,10 +11,10 @@ import ads from "../../../logo.svg";
 import { Link as RouterLink, useRouteMatch } from "react-router-dom";
 import { useIsAuthenticated } from "auth/hooks/queries/useIsAuthenticated";
 import { useSignOut } from "auth/hooks/mutations/useSignOut";
-import { SupportMenu } from "components/Drawer/MiniSideBar";
 import { useIsMobile } from "hooks/useIsMobile";
 import { msg, Trans } from "@lingui/macro";
 import { useLingui } from "@lingui/react";
+import { SupportMenu } from "components/Drawer/SupportMenu";
 
 export function LandingPageAppBar() {
   const match = useRouteMatch();
