+++ src/components/Creatives/CreativeType.tsx
@@ -2,8 +2,8 @@ import { Box, ListItemButton, List, Typography, Stack } from "@mui/material";
 import { useFormikContext } from "formik";
 import { CreativeInput } from "graphql/types";
 import { FormatHelp } from "components/Button/FormatHelp";
-import { msg, Trans } from "@lingui/macro";
-import { Trans as TransWithId } from "@lingui/react";
+import { msg } from "@lingui/macro";
+import { Trans } from "@lingui/react";
 
 export function CreativeType(props: { allowTypeChange?: boolean }) {
   const formik = useFormikContext<CreativeInput>();
@@ -23,7 +23,7 @@ export function CreativeType(props: { allowTypeChange?: boolean }) {
     <Box maxWidth={500} display="flex" flexDirection="column" mt={1}>
       <Stack direction="row" alignItems="center" spacing={1}>
         <Typography>
-          <Trans>Ad Format</Trans>
+          <Trans id={msg`Ad format`.id} />
         </Typography>
         <FormatHelp />
       </Stack>
@@ -43,7 +43,7 @@ export function CreativeType(props: { allowTypeChange?: boolean }) {
             }}
             onClick={() => formik.setFieldValue("type.code", s.value)}
           >
-            <TransWithId id={s.label.id} />
+            <Trans id={s.label.id} />
           </ListItemButton>
         ))}
       </List>
