+++ src/user/analytics/analyticsOverview/components/BasePieChart.tsx
@@ -3,7 +3,7 @@ import { Options, SeriesOptionsType } from "highcharts";
 import { Option } from "../types";
 import { HighchartsWrapper } from "user/analytics/analyticsOverview/components/HighchartsWrapper";
 import { msg } from "@lingui/macro";
-import { Trans as TransWithId } from "@lingui/react";
+import { Trans } from "@lingui/react";
 
 interface Props {
   series: SeriesOptionsType[];
@@ -71,7 +71,7 @@ export function BasePieChart({ series, onSetType, extraOptions, type }: Props) {
             <Tab
               key={t.label.id}
               value={t.value}
-              label={<TransWithId id={t.label.id} />}
+              label={<Trans id={t.label.id} />}
             />
           ))}
         </Tabs>
