+++ src/user/views/adsManager/views/advanced/components/review/components/ReviewField.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography } from "@mui/material";
 import ErrorOutlineIcon from "@mui/icons-material/ErrorOutline";
 import { MessageDescriptor } from "@lingui/core";
-import { Trans as TransWithId } from "@lingui/react";
+import { Trans } from "@lingui/react";
 
 export function DisplayError({ error }: { error: string | undefined }) {
   if (!error) return null;
@@ -24,7 +24,7 @@ export function ReviewField({ caption, value, error }: Props) {
   return (
     <Box mb={2}>
       <Typography variant="overline" component="span" paddingRight={1}>
-        <TransWithId id={caption.id} />
+        <Trans id={caption.id} />
       </Typography>
       <Typography variant="body1" component="span">
         {value}
